backend/requirements.txt
langchain-groq==0.3.2
langchain==0.3.25
pydantic==2.11.5
python-dotenv==1.0.1
pandas==2.2.1
plotly==6.1.0
matplotlib==3.9.4
numpy==1.26.4
sqlalchemy==2.0.28
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.5
groq==0.26.0
seaborn==0.13.2
openpyxl==3.1.2 